Fix issue where HTTPHeaders may allow duplicate headers
It looked like we were trying to use a `TreeSet` with a case insensitive order to avoid duplicate headers.
However when constructing `HTTPHeaders` with a source header map, we were generating the raw headers at the same time as inserting into this map.  Which means we had the potential to be provided duplicate headers which were not recognized.

The fix to this is to iterate over the incoming map twice.  Once to ensure that there is no duplicates (ie build the TreeMap), and a second time to build the raw header string.
Other possibilities:
* We could wait to build the rawHeaders, doing it lazily once it's needed.
* We could as we iterate and insert, look for duplicates and instead error out if any do.  I was thinking that checking while inserting is likely more computationally expensive however, so iterating twice may be the lesser of two evils.