HttpClient: Fix for not timing out queued requests
We would not watch the future until we started processing the request.  This means that if we never started processing the request it would never timeout.
This could have been so bad that eventually we might request a watchdog with a negative timeout.
As an added bonus this reduces the variation of timeouts we may request from the WatchdogCache, allowing it to be much more efficent.