HTTPClient: Allow request data to be streamed easier
This updates HTTPRequestBuilder / ClientHTTPRequest to provide the body as stateful sections.
The entire body can be specified if available still, but it can also be provided as a stream of ListenableFuture<ByteBuffer>'s
Once the body stream is completed the ListenableFuture should provide a null or empty ByteBuffer.
This is a natural transition for chunked encoding (which is now supported in HTTPClient), but can also work for a large streamed body.