HTTPClient: Change how response bodies are handled
This is an API breaking change to change how response bodies are handled.
Instead of having the HTTPClient handle the body, the body handling is instead put into the request (so each request may handle the body differently if needed).
This breaks the API in that HTTPClient is no longer monitoring the maximum body accepted, but rather the default behavior in HTTPRequestBuilder is to buffer and monitor the body limit that was previously done in HTTPClient.