HTTPClient: NullPointerException fix in consuming post bodies
The body provider may provide the body chunks off the client thread.  This means that the client may transition to null if the SettableListenableFuture is completed in error.
If this happens a NullPointerException would have been thrown on line 402 (old code, 411 new code) when the client is attempted to be used for the next write.